// web/src/interfaces/database/agent.ts
export type DSLComponents = Record<string, IOperator>;

export interface DSL {
  components: DSLComponents;
  history: any[];
  path?: string[];
  answer?: any[];
  graph?: IGraph;
  messages?: Message[];
  reference?: IReference[];
  globals: Record<string, any>;
  variables: Record<string, GlobalVariableType>;
  retrieval: IReference[];
}
// ...
export interface IFlowTemplate {
  avatar: string;
  canvas_type: string;
  canvas_types?: string[];
  create_date: string;
  create_time: number;
  canvas_category?: string;
  dsl: DSL;
  id: string;
  update_date: string;
  update_time: number;
  description: {
    en: string;
    zh: string;
    de: string;
  };
  title: {
    en: string;
    zh: string;
    de: string;
  };
}
// ...
export interface IGraph {
  nodes: RAGFlowNodeType[];
  edges: Edge[];
}
// ...
export interface GlobalVariableType {
  name: string;
  value: any;
  description: string;
  type: string;
}